# swagger.json
{
  "swagger": "2.0",
  "title": "Backend in Numerical Project",
  "version": "1.0.0",
  "description": "Backend Numerical",
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Root of Equations"
    },
    {
      "name": "Linear Algebra Equations"
    },
    {
      "name": "Interpolation And Extrapolation"
    },
    {
      "name": "Regression"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/testpost": {
      "post": {
        "tags": [
          "Root of Equations"
        ],
        "description": "Bisection Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testfalse": {
      "post": {
        "tags": [
          "Root of Equations"
        ],
        "description": "False-Position Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testonepoint": {
      "post": {
        "tags": [
          "Root of Equations"
        ],
        "description": "One-Point Iteration Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testcramer": {
      "post": {
        "tags": [
          "Linear Algebra Equations"
        ],
        "description": "Cramer's Rule",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },

    "/testNewtons2": {
      "post": {
        "tags": [
          "Interpolation And Extrapolation"
        ],
        "description": "Newton's ",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testlagrange": {
      "post": {
        "tags": [
          "Interpolation And Extrapolation"
        ],
        "description": "Lagrange Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testspline": {
      "post": {
        "tags": [
          "Interpolation And Extrapolation"
        ],
        "description": "Spline Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testLine": {
      "post": {
        "tags": [
          "Regression"
        ],
        "description": "Linear Regression Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testPoly": {
      "post": {
        "tags": [
          "Regression"
        ],
        "description": "Polynomial Regression Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/testMaltiple": {
      "post": {
        "tags": [
          "Regression"
        ],
        "description": "Multi Linear Regression Method",
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  }
}